% Line Tessellation (LiTe) library
% Release 1.1
% Authors: Katarzyna Adamczyk and Kiên Kiêu.
% Copyright INRA 2006-2015.
% Interdeposit Certification: IDDN.FR.001.030007.000.R.P.2015.000.31235
% License: GPL v3.
\name{Rcpp_SMFChain-class}
\docType{class}
\alias{Rcpp_SMFChain-class}
\title{S4 Gibbsian T-tessellation simulator}
\description{S4 class associated with \code{\link{SMFChain}} objects.}
\section{S4 Methods}{
  No specific S4 method has been defined for this class.
  }